+++ src/main/java/fr/bmj/bmjc/dataaccess/DataAccessManagePlayer.java
@@ -26,8 +26,8 @@ public interface DataAccessManagePlayer extends DataAccessCommon {
 
 	public List<Player> getAllPlayers();
 
-	public UpdateResult modifyPlayer(final int id, String name, final String displayName, boolean frequent, boolean regular, String license);
+	public UpdateResult modifyPlayer(final short id, String name, final String displayName, boolean frequent, boolean regular, String license);
 
-	public UpdateResult deletePlayer(final int id);
+	public UpdateResult deletePlayer(final short id);
 
 }
